Fixed compiler warning

+++ frontend_local.c
@@ -4,7 +4,7 @@
  * See the main source file 'xineliboutput.c' for copyright information and
  * how to reach the author.
  *
- * $Id: frontend_local.c,v 1.22 2007-03-14 11:44:30 phintuka Exp $
+ * $Id: frontend_local.c,v 1.23 2007-03-14 13:21:08 phintuka Exp $
  *
  */
 
@@ -206,7 +206,7 @@ int cXinelibLocal::Xine_Control(const char *cmd)
   TRACEF("cXinelibLocal::Xine_Control");
   if(cmd && *cmd && !GetStopSignal()) {
     char buf[4096];
-    if(snprintf(buf, sizeof(buf), "%s\r\n", cmd) >= sizeof(buf)) {
+    if(snprintf(buf, sizeof(buf), "%s\r\n", cmd) >= (int)sizeof(buf)) {
       buf[sizeof(buf)-1] = 0;
       LOGMSG("Xine_Control: message too long ! (%s)", buf);
       return 0;
